refactor: move style.ts to style folder
