+++ src/shared/styles/index.ts
@@ -1,4 +1,3 @@
-// import { Colors } from 'react-native/Libraries/NewAppScreen';
 import { StyleSheet } from 'react-native';
 
 export const fontStyle = StyleSheet.create({
